Доделал локализацию без переключения

+++ lib/core/chats/gpt/services/chat_gpt_service.dart
@@ -172,7 +172,7 @@ class ChatGptService {
 
     getIt<Talker>().info('EXPITATION DAYS = ${expiration}');
 
-    if (expiration != null) {
+    if (expiration != null && expiration != -1) {
       isDeleted = await messageRepository.deleteMessagesOlderThan(expiration);
     }
